requirements.txt
streamlit
pandas
python-docx
docx2pdf
smtplib
